ft_strlen updated

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: ztaskula <ztaskula@student.42istanbul.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/18 21:16:45 by zzehra            #+#    #+#             */
-/*   Updated: 2025/06/28 14:38:48 by ztaskula         ###   ########.fr       */
+/*   Updated: 2025/06/30 15:59:12 by ztaskula         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	while (str[len])
